+++ src/template.js
@@ -1,0 +1,8 @@
+class Template {
+  constructor(description, completed, index) {
+    this.description = description;
+    this.completed = completed;
+    this.index = index;
+  }
+}
+export default Template;
